Modified the class to be unit testable.
