+++ src/CaptainHook.DirectorService/Program.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Diagnostics;
 using System.Fabric;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Autofac;
@@ -17,7 +18,16 @@ using CaptainHook.DirectorService.ReaderServiceManagement;
 using CaptainHook.Storage.Cosmos;
 using Eshopworld.Data.CosmosDb.Extensions;
 using Eshopworld.Telemetry;
+using Microsoft.Azure.Management.ResourceManager.Fluent;
+using Microsoft.Azure.Management.ResourceManager.Fluent.Authentication;
+using Microsoft.Azure.Management.ResourceManager.Fluent.Core;
+using Microsoft.Azure.Management.ServiceBus.Fluent;
+using Microsoft.Azure.Services.AppAuthentication;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Rest;
+using Nito.AsyncEx;
+using IServiceBusManager = CaptainHook.Common.ServiceBus.IServiceBusManager;
+using ServiceBusManager = CaptainHook.Common.ServiceBus.ServiceBusManager;
 
 namespace CaptainHook.DirectorService
 {
@@ -87,6 +97,8 @@ namespace CaptainHook.DirectorService
                     .As<IMessageProviderFactory>()
                     .SingleInstance();
 
+                builder.ConfigureServiceBusManager(configurationSettings);
+
                 builder.RegisterType<ServiceBusManager>()
                     .As<IServiceBusManager>();
 
@@ -146,5 +158,7 @@ namespace CaptainHook.DirectorService
 
             return value;
         }
+
+        
     }
 }
